requirements.txt
Django
whitenoise
azure-monitor-opentelemetry